+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "hubot-hockey",
   "description": "Get the latest NHL playoff odds for your team.",
-  "version": "6.5.1",
+  "version": "6.6.0",
   "author": "Stephen Yeargin <stephen@yearg.in>",
   "homepage": "https://github.com/stephenyeargin/hubot-hockey",
   "license": "MIT",
@@ -28,18 +28,14 @@
     "hubot": ">=3 || 0.0.0-development"
   },
   "devDependencies": {
-    "chai": "^4.3.10",
-    "eslint": "^8.54.0",
-    "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-plugin-import": "^2.32.0",
-    "eslint-plugin-n": "^16.3.1",
-    "eslint-plugin-promise": "^6.1.1",
+    "chai": "^6.2.0",
+    "eslint": "^9.38.0",
     "hubot-mock-adapter": "^2.3.0",
     "husky": "^9.1.7",
-    "mocha": "^10.2.0",
-    "nock": "^14.0.8",
-    "sinon": "^17.0.1",
-    "sinon-chai": "^3.7.0"
+    "mocha": "^11.7.4",
+    "nock": "^14.0.10",
+    "sinon": "^21.0.0",
+    "sinon-chai": "^4.0.1"
   },
   "main": "index.js",
   "scripts": {
