+++ test/hockey-slack_test.mjs
@@ -1,13 +1,14 @@
-/* eslint-disable no-undef */
+import hubot from 'hubot';
+import { expect } from 'chai';
+import nock from 'nock';
+import path from 'node:path';
+import { fileURLToPath } from 'node:url';
 
-const { Robot, User, TextMessage } = require('hubot');
-const chai = require('chai');
-const nock = require('nock');
-const path = require('path');
+const { Robot, User, TextMessage } = hubot;
 
-const {
-  expect,
-} = chai;
+// ESM-friendly __dirname
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
 
 // Alter time as test runs
 const originalDateNow = Date.now;
@@ -1104,63 +1105,6 @@ describe('hubot-hockey for slack', () => {
                        + '```',
                     title: '4/23/2024 - Final - R1 Game 2 (Tied 1-1)',
                     title_link: 'https://www.nhl.com/gamecenter/2023030172',
-                  },
-                ],
-              },
-            ],
-          ]);
-          done();
-        } catch (err) {
-          done(err);
-        }
-      },
-      100,
-    );
-  });
-
-  it('responds with a completed playoff series', (done) => {
-    Date.now = () => Date.parse('Fri May 3 22:00:00 CST 2024');
-    nock('https://api-web.nhle.com')
-      .get('/v1/scoreboard/nsh/now')
-      .replyWithFile(200, `${__dirname}/fixtures/api-web-nhle-schedule-eliminated.json`);
-
-    nock('https://moneypuck.com')
-      .get('/moneypuck/simulations/update_date.txt')
-      .reply(200, '2023-05-03 06:52:52.999000-04:00');
-
-    nock('https://moneypuck.com')
-      .get('/moneypuck/simulations/simulations_recent.csv')
-      .replyWithFile(200, `${__dirname}/fixtures/moneypuck-simulations_recent.csv`);
-
-    // Using userSays helper
-    userSays('alice', '@hubot preds');
-    setTimeout(
-      () => {
-        try {
-          expect(messages).to.eql([
-            ['alice', '@hubot preds'],
-            [
-              'hubot',
-              {
-                attachments: [
-                  {
-                    author_icon: 'https://github.com/nhl.png',
-                    author_link: 'https://nhl.com',
-                    author_name: 'NHL.com',
-                    color: '#FFB81C',
-                    fallback: '5/3/2024 - Vancouver Canucks (9-2-1) 1, Nashville Predators (5-6-0) 0 (Final - R1 Game 6 (VAN wins 4-2))',
-                    footer: 'Bridgestone Arena',
-                    mrkdwn_in: [
-                      'text',
-                      'pretext',
-                    ],
-                    text:
-                      '```\n'
-                       + '  Vancouver Canucks (9-2-1)     1  \n'
-                       + '  Nashville Predators (5-6-0)   0  \n'
-                       + '```',
-                    title: '5/3/2024 - Final - R1 Game 6 (VAN wins 4-2)',
-                    title_link: 'https://www.nhl.com/gamecenter/2023030176',
                   },
                 ],
               },
